Rename files in /lib
