+++ lib/api.js
@@ -2,10 +2,10 @@
 
 const path = require('path');
 const fs = require('mz/fs');
+const debug = require('debug')('madge');
 const dependencyTree = require('dependency-tree');
 const cyclic = require('./cyclic');
 const graph = require('./graph');
-const debug = require('debug')('madge');
 
 const defaultConfig = {
 	showFileExtension: false,
