+++ lib/output.js
@@ -110,3 +110,13 @@ module.exports.depends = function (modules, opts) {
 		console.log(c(id, 'grey', opts.colors));
 	});
 };
+
+/**
+ * Print error to the console.
+ * @param  {Object} err
+ * @param  {Object} opts
+ * @return {undefined}
+ */
+module.exports.error = function (err, opts) {
+	console.log(c(err, 'red', opts.colors));
+};
