+++ package.json
@@ -22,6 +22,10 @@
   "engines": {
     "node": ">=4.x.x"
   },
+  "main": "./lib/api",
+  "bin": {
+    "madge": "./bin/cli.js"
+  },
   "scripts": {
     "test": "npm run lint && npm run madge && npm run mocha",
     "mocha": "mocha test/*.js",
@@ -47,9 +51,5 @@
     "mocha": "^2.3.3",
     "release-it": "^2.4.0",
     "should": "^9.0.2"
-  },
-  "main": "./lib/madge",
-  "bin": {
-    "madge": "./bin/cli.js"
   }
 }
